% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{bestMclust}
\alias{bestMclust}
\title{Get best cluster for a given number of best clusters to search
Credits to the developers of the edci package in R}
\usage{
bestMclust(clust, nc = 1, crit = "value")
}
\arguments{
\item{clust}{cluster object as from oregMclust or circMclust}

\item{nc}{number of best cluster}

\item{crit}{which column extracts best clusters, default: "value"}
}
\description{
Get best cluster for a given number of best clusters to search
Credits to the developers of the edci package in R
}
