% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{centrit}
\alias{centrit}
\title{Convert points into polygon for circle section by ordering to angle - not used in package}
\usage{
centrit(pts)
}
\arguments{
\item{pts}{spatial points object}
}
\description{
Convert points into polygon for circle section by ordering to angle - not used in package
}
