% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2_instanceSegmentation.R
\name{changeLASVEG}
\alias{changeLASVEG}
\title{Report current LAS_veg name and also outLAS name (to quickly do segmentation without reading in laz)}
\usage{
changeLASVEG()
}
\description{
Report current LAS_veg name and also outLAS name (to quickly do segmentation without reading in laz)
}
