% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{circMclust}
\alias{circMclust}
\title{Debugged function to get circular clusters from edci
for parameter help please refer to
https://rdrr.io/cran/edci/man/circMclust.html}
\usage{
circMclust(
  datax,
  datay,
  bw,
  method = "const",
  prec = 4,
  minsx = min(datax),
  maxsx = max(datax),
  nx = 10,
  minsy = min(datay),
  maxsy = max(datay),
  ny = 10,
  minsr = 0.01 * max(datax, datay),
  maxsr = (max(datax, datay) - min(datax, datay)),
  nr = 10,
  nsc = 5,
  nc = NULL,
  minsd = NULL,
  maxsd = NULL,
  brminx = minsx,
  brmaxx = maxsx,
  brminy = minsy,
  brmaxy = maxsy,
  brminr = minsr,
  brmaxr = maxsr,
  brmaxit = 1000
)
}
\description{
Debugged function to get circular clusters from edci
for parameter help please refer to
https://rdrr.io/cran/edci/man/circMclust.html
}
