% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x1_treeLocalisationDensityClusters.R
\name{clustSplit}
\alias{clustSplit}
\title{Individual tree detection}
\usage{
clustSplit(
  fileFinder,
  allDBHs = FALSE,
  allFiles = FALSE,
  clipHeight = 3,
  bottomCut = 1,
  ipad = FALSE,
  nr_cores = 0,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  filterINT = 0,
  ref = NA,
  ref.plot_id = NA,
  cutWindow = c(-1000, -1000, 2000),
  numberOfPoints = 300,
  heightExtent = 1.3,
  TLS = FALSE,
  silent = TRUE,
  fast = TRUE,
  retainPointClouds = TRUE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{allDBHs}{add all measured diameters (gam, circle, ellipsis) to the output file trees_dbh.txt}
}
\description{
Reads ground and vegetation point cloud and performs
tree localisation by density cluster analysis and
produces trees_dbh.txt file
}
