% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3_calculateTreeParameters.R
\name{computeCrownParams}
\alias{computeCrownParams}
\title{Calculation of individual tree measurements}
\usage{
computeCrownParams(
  fileFinder,
  loopStart = 1,
  loopEnd = 0,
  drawImages = T,
  remeasureDBH = F,
  decimateTreeForFasterCrowns = FALSE,
  mode = "ALLGO",
  maxRadius = 0,
  ipad = FALSE,
  cutWindow = c(-1000, -1000, 2000),
  zScale = 2,
  limitShare = 0.004,
  voxelSize = 0,
  vol.alpha = 2,
  alternativeCrownBase.Ratio = 0.3,
  fogFilter.estHeight = FALSE,
  detail.level = 0,
  selector = "xyzcit0",
  limitSpanSide = 30,
  limitSpanArea = 440,
  referenceDiameterLimit = 150,
  quantileIntensity = 15,
  numberIntensity = 0,
  CC_level = 10,
  CC_numberPoints = 1000,
  clipHeight = 3,
  bottomCut = 1,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  do.CrownBase = TRUE,
  groundCutHeight = 0,
  silent = TRUE,
  retainPointClouds = FALSE,
  nr_cores = 0,
  treeNames = NULL,
  smallTrees = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\description{
Input files are  single tree .laz files, there is no splitting of
the segmented crowns file.
}
