% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3_calculateTreeParameters.R
\name{computeTree_i}
\alias{computeTree_i}
\title{calculates the individual tree parameters for dedicated tree}
\usage{
computeTree_i(
  treeLAS.path,
  treeName = "",
  drawImage = F,
  z_stemBase = NA,
  nowMeta = NA,
  detail.level = 0,
  decimateTreeForFasterCrowns = FALSE,
  vol.alpha = 2,
  area.alpha = 0.3,
  alternativeCrownBase.Ratio = 0.3,
  fogFilter.estHeight = FALSE,
  measurePath = dirname(treeLAS.path),
  fileFinder = "",
  remeasureDBH = FALSE,
  do.Plot = TRUE,
  limitSpanSide = 30,
  limitSpanArea = 440,
  referenceDiameterLimit = 150
)
}
\arguments{
\item{treeLAS.path}{individual Tree LAS file}

\item{treeName}{additional name for the images (specific trees)}

\item{z_stemBase}{predefined z-coordinate for start of the stem, overrides nowMeta$z}

\item{nowMeta}{additional information about tree species, xyz coordinates and dbh}

\item{detail.level}{how many points shall be kept of crown file for hulling 0 = standard (2 pts every 25x25cm), -1 = rough (5 pts per 1x1m), -3 very rough (5 pts per 3x3m), 3 very fine (3 pts every 10x10cm)
TODO: implement the crownBase list input with passing the column crownBase in nowMeta, I think I fixed it in 23-11 (without testing)}

\item{tree}{number of StemID which to measure}
}
\description{
calculates the individual tree parameters for dedicated tree
}
