% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3_calculateTreeParameters.R
\name{computeTreeParams}
\alias{computeTreeParams}
\title{Calculation of individual tree measurements}
\usage{
computeTreeParams(
  fileFinder,
  loopStart = 1,
  loopEnd = 0,
  getRAM = FALSE,
  detail.level = 0,
  drawNumberPoints = TRUE,
  writeLAZ = TRUE,
  writePicture = TRUE,
  crownParameters = TRUE,
  mode = "ALLGO",
  maxRadius = 0,
  ipad = FALSE,
  cutWindow = c(-1000, -1000, 2000),
  zScale = 2,
  limitShare = 0.004,
  voxelSize = 0,
  vol.alpha = 2,
  alternativeCrownBase.Ratio = 0.3,
  fogFilter.estHeight = FALSE,
  selector = "xyzcit0",
  limitSpanSide = 30,
  limitSpanArea = 440,
  referenceDiameterLimit = 150,
  quantileIntensity = 15,
  numberIntensity = 0,
  CC_level = 10,
  CC_numberPoints = 1000,
  clipHeight = 3,
  bottomCut = 1,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  do.CrownBase = TRUE,
  groundCutHeight = 0,
  silent = TRUE,
  retainPointClouds = FALSE,
  nr_cores = 0,
  nr_cores_params = 0,
  treeNames = NULL,
  smallTrees = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{writeLAZ}{export the single trees as .laz}

\item{writePicture}{export the single trees as front, side and top view image, INCREASES COMPUTATION TIME}

\item{maxRadius}{for Ebensee implemented, in a circle of 20 m all trees will be exported, but not the outer trees}

\item{vol.alpha}{defines the alpha hull of crown volume, if 2 very rough (balloons) if 0.4 very tight (old settings)}

\item{fogFilter.estHeight}{if TRUE the height will be calculated to 99.9 \% quantile to the highest point, if FALSE to max height of the tree}

\item{groundCutHeight}{how much distance between lowest point of tree and ground (if ground was cut off)}

\item{fileName}{origin of the segmented crown las file in dirPath}
}
\description{
Loads in all single seperated Tree files from segmented _crowns.laz
and calculates for each individual the crown specific parameters.
E.g.: tree height, crown base, crown volume, crown projection area, ...
makes .png pictures and saves the list as
trees_height.txt and trees_measured.txt (incl. crown parameters)
}
