% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R
\name{createAppFiles}
\alias{createAppFiles}
\title{Creation of APP files (background image and tree list)}
\usage{
createAppFiles(
  fileFinder = NA,
  drawGround = T,
  fullGroundGrey = F,
  drawRedSlice = T,
  pixelUnit_cm = 2,
  createBGR_pic = TRUE,
  createJPG = TRUE,
  jpgQuality = 0.3,
  createTIFF = FALSE,
  dirPath = paste0(getwd(), "/"),
  changeFileFinder = "",
  addTrajLAZ = F,
  drawTraj = FALSE,
  greySpots = data.frame(),
  drawLines = data.frame(),
  cutWindow = c(-1000, -1000, 2000, 2000),
  reduceName = FALSE,
  onlySuffixFileFinder = FALSE,
  slices = c("_clusterSlice_100to300.laz"),
  laz.path = NA,
  slice.cut.low = NA,
  slice.cut.high = NA,
  colorVersion = F,
  eraseSpecies = F,
  highlightNew = F,
  setSpecies = "",
  dtm.path = "",
  trees.path = "",
  isoLines = 0,
  thin = FALSE,
  modeApp = 1,
  selector = "xyzi",
  writeColoredLAZ = FALSE,
  circleRadius = 0,
  fixedLimit = 0,
  wait5s = FALSE,
  quickPreview = F,
  greenTreeLocations = F
)
}
\arguments{
\item{fileFinder}{user defined name to find this certain dataset in further processing}

\item{drawGround}{set FALSE to omitt the voxelized ground points lower than 1 m}

\item{fullGroundGrey}{set TRUE to plot full ground points from 1 m lower in grey (instead of voxelized points)}

\item{pixelUnit_cm}{how many cms are represented in one pixel - default 1 (100 pixel equals one meter), also 2 works for smaller plots}

\item{changeFileFinder}{rename app background to this name}

\item{circleRadius}{radius of the additional circle that is plot in blue on top of the background file - recommended: 5 m more than desired sample plot radius}

\item{fixedLimit}{limit in meters that defines boundary of background file - for round circles set it to the clipped file size, so that all backgrounds have the same dimensions}

\item{quickPreview}{set TRUE to plot only reduced density of slice}
}
\description{
Reads in the las slices of the given set in fileFinder,
and uses trees_dbh.txt or trees_measured.txt to export the tree list
}
