% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2_instanceSegmentation.R
\name{crownFeel}
\alias{crownFeel}
\title{Tree crown segmentation}
\usage{
crownFeel(
  fileFinder,
  cutWindow = c(-1000, -1000, 2000),
  ipad = FALSE,
  doReferencedOnly = FALSE,
  referenced = FALSE,
  zScale = 2,
  merged = FALSE,
  totalRuns = 1000,
  limitShare = 0.004,
  incrementDistance = 0.005,
  limitStems = 0,
  tileClipping = 5,
  diagonals = FALSE,
  useTreeFile = "",
  selector = "xyzcit0",
  distanceCounter_cm_limit = 220,
  quantileIntensity = 15,
  CC_level = 10,
  CC_numberPoints = 1000,
  durationMins = 2400,
  maximumDistance = 0.5,
  clipHeight = 3,
  bottomCut = 1,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  voxelSize = 0,
  numberIntensity = 0,
  silent = TRUE,
  fast = TRUE,
  plotEvery = 100,
  mode = "ALLGO",
  frame.up = 0.3,
  frame.down = 0.3,
  frame.rad = 1.05,
  retainPointClouds = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{fileFinder}{user defined name of the dataset}

\item{cutWindow}{c(xL, yL, width) square frame to detect trees from}

\item{doReferencedOnly}{deletes all the above 20000 stems, they do not make senseful trees}

\item{zScale}{relative number of Z-variable multiplication for a better height progress}

\item{merged}{set TRUE if using a combined stem-file with field "size" for combining}

\item{totalRuns}{number of iterations for discovering crowns, default 500}

\item{limitShare}{if we are adding less points from blanks than that, then incrementing searching distance (default: 0.005 = 0.5 \%)}

\item{incrementDistance}{the step of raising searching distance if less than limitShare points are found}

\item{useTreeFile}{loading and external trees_dbh.txt file}

\item{CC_level}{a connected components parameter for fineness in cloudcompare (default: 10)}

\item{CC_numberPoints}{the minimum size in points of a component to not be dropped (default: 1000)}

\item{clipHeight}{in m if you need to set all stem seeds to a uniform height}

\item{bottomCut}{only in combination with clipHeight, in m lower border for cutting of relative heights, default to 0}

\item{bushPreparation}{if TRUE then there is additional density filtering to remove low branches and bushes}

\item{filterSOR}{if TRUE then we use the SOR clean file}

\item{voxelSize}{before the region growing the point cloud can be voxelized, in cm, default = 0 (no voxelisation)}

\item{numberIntensity}{if you rather want to specify a number of most intense points to be kept}

\item{silent}{fewer outputs or plots if set to TRUE}

\item{fast}{if TRUE, no detailled per stem information will be calculated}

\item{retainPointClouds}{if TRUE, the used point clouds will be stored for another run (more performant for different settings), if FALSE the point clouds will be passed to next function layer (always) and then discarded.}
}
\description{
Classifies all trees and isolates the crown,
stem coordinates are required from trees_dbh.txt
which are used as seed points, from there it grows the stems
until all points are allocated to a stemID
}
