% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{deldupMclust}
\alias{deldupMclust}
\title{Delete duplicated clusters via rounding errors or maximization value of zero
Credits to the developers of the edci package in R}
\usage{
deldupMclust(clust, prec = NULL, ncol = NULL, dz = TRUE)
}
\arguments{
\item{clust}{numerical matrix where columns are parameters of clusters}

\item{prec}{decimal places for rounding, positive integer, optional}

\item{ncol}{number of columns that gives clusters}

\item{dz}{when TRUE, the clusters with 0 are deleted}
}
\description{
Delete duplicated clusters via rounding errors or maximization value of zero
Credits to the developers of the edci package in R
}
