% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x0_groundRemovalDTM.R
\name{extractVegetation}
\alias{extractVegetation}
\title{Ground classification of raw point cloud}
\usage{
extractVegetation(
  LASfile,
  fileFinder,
  groundMergeCut = 0,
  ipad = FALSE,
  groundCutHeight = 1,
  steepSlope = TRUE,
  clothSize = 0.1,
  doFilter = FALSE,
  filterRes = 0.05,
  filterN = 27,
  tooBig = F,
  clip.radius = 0,
  clip.trajectory.distance = 0,
  clip.x = 0,
  clip.y = 0,
  selector = "xyzcit0",
  dtm.path = "",
  trafoMatrix.path = "",
  rainFilter = 0,
  additionalSlices = TRUE,
  exportSlice.upperLimit = 3,
  exportSlice.lowerLimit = 1,
  exportClippedLAS = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{LASfile}{input path of full (or cropped) geoslam .las data}

\item{fileFinder}{user defined name to find this certain dataset in further processing}

\item{groundMergeCut}{in cm, how high should additional slice from ground be for a merging dtm (multiple tiles)}

\item{groundCutHeight}{in m, lower than this is all classified as ground (default: 0.5 m)}

\item{steepSlope}{in mountainous areas set it TRUE for better ground model}

\item{clothSize}{resolution of the cloth grid}

\item{filterRes}{noise filter resolution for adjacent 3x3x3 = 27 voxels, 2 cm (strong filter) takes longer time than 5 cm (medium filter)}

\item{filterN}{number of neighbour points, "strength" of filter, how many neighbours there need to be in all 27 neighbour voxels, (8 would be a medium, 27 a strong filter)}

\item{clip.radius}{can be set to cut a circle from 0 0 with the given radius of the big input file}

\item{clip.trajectory.distance}{can be set to cut the input file according to the trajectory}

\item{additionalSlices}{should additional slices be saved as .laz files}

\item{exportSlice.upperLimit}{in m, upper limit of single slice exported after ground classification for previewing stems and tree detection, set to 0 to disable slice export}

\item{exportSlice.lowerLimit}{in m, lower limit of single exported slice, default = 1 m}

\item{exportClippedLAS}{if set TRUE, then a laz file will be exported according to clip.radius}
}
\description{
Reads in a .las or .laz file from disk and extracts the vegetation cloud
using the cloth simulation filter (scf) via lasground from the lidR package
}
