% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x1_treeLocalisationDensityClusters.R
\name{fineCluster_i}
\alias{fineCluster_i}
\title{analyzes, if the measured cluster can be a tree}
\usage{
fineCluster_i(circleFile)
}
\arguments{
\item{circleFile}{name of the file to be analyzed}
}
\description{
analyzes, if the measured cluster can be a tree
}
