% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R
\name{fPoll}
\alias{fPoll}
\title{Calculate form factor (Formzahl) according to Pollanschuetz given species, dbh and height}
\usage{
fPoll(species, dbh_cm, height_m)
}
\arguments{
\item{species}{number for species from function treeSpeciesNumber, spruce = 1}

\item{dbh_cm}{diameter at breast height in cm}

\item{height_m}{tree height in meters}
}
\description{
Calculate form factor (Formzahl) according to Pollanschuetz given species, dbh and height
}
