% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{generateSetString}
\alias{generateSetString}
\title{Separate function to generate the addressing string to store the files with all the settings in the name}
\usage{
generateSetString(
  fileFinder = NA,
  mode = NA,
  threshold = NA,
  threshold_percent = NA,
  bottomCut = NA,
  clipHeight = 0,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  level = NA,
  numberOfPoints = NA,
  cutWindow = c(-1000, -1000, 2000),
  limitShare = NA,
  zScale = NA,
  voxelSize = 0,
  slices = NA,
  bottom = NA,
  step = NA,
  overlap = NA,
  referenced = FALSE,
  silent = FALSE
)
}
\description{
Separate function to generate the addressing string to store the files with all the settings in the name
}
