% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad2_remeasureNewTrees.R
\name{grabDBH}
\alias{grabDBH}
\title{Remeasure best fitting DBH for new trees that have xy and approximate DBH}
\usage{
grabDBH(
  fileFinder,
  treeList.path = NA,
  transformList.fileFinder = "",
  transformList.path = "",
  transformList.matrix = "",
  transformSlice = "_clusterSlice_120to140.laz",
  moveOnly = F,
  dirPath = paste0(getwd(), "/"),
  tileClipping = 3,
  overWriteDBHlist = TRUE,
  keepBorderTrees = F,
  remeasure = T,
  allTrees = F,
  new.numbers = F,
  frame.up = 0.3,
  frame.down = 0.3,
  frame.rad = 1.3,
  frame.minD = 0.05
)
}
\arguments{
\item{treeList.path}{input path of trees that should be re-measured}

\item{moveOnly}{set to TRUE to keep old DBHs in the main list (in case there is a lot of movement expected and circles are not perfectly centered)}

\item{tileClipping}{how many tiles should be created (default 3 means: 3x3 = 9 tiles), set to 4 (= 16 tiles) for big files}

\item{remeasure}{creates new cylinderLAS for measuring diameters, set FALSE only if cylinderLAS is already up to date, otherwise some trees might be missing}

\item{allTrees}{set this TRUE to also measure trees with numbers less than 9000}

\item{new.numbers}{assign new sequential numbers for circles (1 being the thickest tree)}

\item{frame.up}{delta z up from 1.30 m for circle fitting}

\item{frame.down}{delta z down from 1.30 m for circle fitting}

\item{frame.rad}{re-locating circles: factor of which input DBH will be multiplied to extend circle-fitting}

\item{frame.minD}{minimum distance (in meter) to fetch the circle around for re-locating, set to 1 m for a tolerant fit (mis-matching systems)}
}
\description{
Remeasure best fitting DBH for new trees that have xy and approximate DBH
}
