% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R
\name{openColoredLAZ}
\alias{openColoredLAZ}
\title{opens multiple height slices in CloudCompare
for better visualization and finding omitted trees}
\usage{
openColoredLAZ(
  fileFinder,
  withGround = T,
  dirPath = paste0(getwd(), "/", ),
  pathCloudCompare = "\\"C:/Program Files/CloudCompare/\\""
)
}
\arguments{
\item{fileFinder}{name of the set to open in dirPath}

\item{withGround}{should ground file (usually quite heavy in data) also be opened? default: TRUE}

\item{dirPath}{directory that contains fileFinder folders and _total_ground_veg}

\item{pathCloudCompare}{path leading to CloudCompare.exe, default for windows in C ProgramFiles}
}
\description{
opens following slices:
\itemize{
\item red DBH (120 - 140 cm)
\item bw 100 - 300 cm
\item bw 300 - 500 cm
(* bw 0 - 100 cm (ground))
' grid 5 x 5 m (5 x 10 m)
}
}
\details{
optionally opens ground (0 - 100 cm) by setting    withGround <- T
}
