% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3_calculateTreeParameters.R
\name{plotOneTree}
\alias{plotOneTree}
\title{Plot one tree from single .laz file}
\usage{
plotOneTree(treePath = "D:/tree0354.laz", stemBase = 12.3)
}
\arguments{
\item{treePath}{path to laz file that is to be plot}

\item{stemBase}{relative z coordinate of bottom of stem (z.DBH - 1.3 m) to get height and stem length correctly}
}
\description{
Plot one tree from single .laz file
}
