% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{processPlotsParallel}
\alias{processPlotsParallel}
\title{Full tree detection and instance segmentation workflow}
\usage{
processPlotsParallel(
  inputFiles,
  fileFinders = "",
  dirPath = paste0(getwd(), "/"),
  maximum_cores = 0,
  nr_plots_parallel = 20,
  nr_cores_per_plot = 1,
  trafoFiles = "",
  groundModels = TRUE,
  detectTrees = TRUE,
  segmentTrees = FALSE,
  crownParameters = TRUE,
  createAppFiles = TRUE,
  setting = "",
  clip.trajectory.distance = 0,
  clip.radius = 0,
  pixelUnit_cm = 2,
  circleRadius = 0,
  drawTraj = F,
  createBGR_pic = T,
  voxelSize = 3,
  limitShare = 0.003,
  limitStems = 50,
  zScale = 3,
  tileClipping = 2
)
}
\arguments{
\item{inputFiles}{vector of paths to .laz or .las files as input for processing (plots)}

\item{fileFinders}{vector of file names for renaming input files. Must be of same length as lazFiles}

\item{maximum_cores}{quick setting to process 20 plots at the same time with 10 cores per plot}

\item{nr_plots_parallel}{how many plots should be calculated simultaneously}

\item{nr_cores_per_plot}{how many cores should each plot use}

\item{trafoFiles}{vector of path to transformation matrices for each plot}

\item{detectTrees}{runs functions extractVegetation and clustSplit}

\item{segmentTrees}{runs functions crownFeel and computeTreeParams}

\item{crownParameters}{also performs stem analysis and extracts crown basal height, projection area and hull volume}

\item{createAppFiles}{runs function createAppFiles with new background images}

\item{setting}{quick preset for 0=ground, 1=trees, 2=segment, a=app, 9=all
#' @export}

\item{circleRadius}{draw additional circle in the background files}
}
\description{
Runs the complete workflow for multiple plots parallel.
Be careful, as this process can be computationally demanding
and the system might fail when limited in RAM.
I recommend having at least 20 GB of RAM for each sample plot radius 20 m
and at least 60 GB of RAM for each scan up to 1 ha.
}
