% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{pullFromFaroConnect}
\alias{pullFromFaroConnect}
\title{Copies (or moves) all laz and trajectory files
from the FARO connect directory to a specified path}
\usage{
pullFromFaroConnect(
  inConnectPath,
  outNewFolder,
  move = F,
  handleCleanLAZ = -1,
  cleanConnectPath = F
)
}
\arguments{
\item{inConnectPath}{path to the input folders of all scans}

\item{outNewFolder}{path where all laz and trajs will be copied (or moved)}

\item{move}{set TRUE to move instead of copy files}

\item{handleCleanLAZ}{what to do with files ending in _clean.laz: -1 (default) ignore them, 0 move them as well, 1 only move _clean.laz files}

\item{cleanConnectPath}{set TRUE to delete all other files in inConnectPath}
}
\description{
Copies (or moves) all laz and trajectory files
from the FARO connect directory to a specified path
}
