% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad2_remeasureNewTrees.R
\name{remeasureNewTrees}
\alias{remeasureNewTrees}
\title{Runs the function grabDBH( ) for all latest files in one directory}
\usage{
remeasureNewTrees(
  dir_completedInputLists,
  appendix = "",
  fileFinders_selected = "",
  dirPath = paste0(getwd(), "/"),
  tileClipping = 2
)
}
\arguments{
\item{dir_completedInputLists}{path to the list of exported trees (new trees > 9000)}

\item{dirPath}{path to all processed files after function extractVegetation( )}

\item{tileClipping}{global setting for all tiles (standard setting 2x2 for small points, if large forest scan, set to 4x4 or 5x5 tiles)}
}
\description{
Runs the function grabDBH( ) for all latest files in one directory
}
