% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x1_treeLocalisationDensityClusters.R
\name{removeUmlaut}
\alias{removeUmlaut}
\title{Remove German Umlaut letters and convert them}
\usage{
removeUmlaut(inputString)
}
\arguments{
\item{inputString}{text that might contain Umlaut letters}
}
\value{
text that is free from Umlaut letters. Throws a warning in case a Umlaut was removed.
}
\description{
Improves handling of ae, oe, ue and ss letters to prevent issues with various encodings
}
