% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R
\name{statsCompletedTrees}
\alias{statsCompletedTrees}
\title{Creates a list of how many new trees were added ( id > 9000 )
and also gives the mean, max and median of new and old DBH}
\usage{
statsCompletedTrees(dir_completedInputLists)
}
\arguments{
\item{dir_completedInputLists}{path to the list of exported trees (new trees > 9000)}
}
\description{
Creates a list of how many new trees were added ( id > 9000 )
and also gives the mean, max and median of new and old DBH
}
