% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x0_groundRemovalDTM.R
\name{stemSplit}
\alias{stemSplit}
\title{Takes the vegetation cloud from "extractVegetation" and seperates the stems
according to their intensity values}
\usage{
stemSplit(
  fileFinder,
  quantileIntensity = 15,
  CC_level = 10,
  CC_numberPoints = 1000,
  clipHeight = 3,
  bottomCut = 1,
  bushPreparation = FALSE,
  filterSOR = FALSE,
  numberIntensity = 0,
  silent = TRUE,
  fast = TRUE,
  groundCutHeight = 0.5,
  amplitudeLowPass = 25.8,
  cutWindow = c(-1000, -1000, 2000),
  retainPointClouds = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{fileFinder}{user defined name of the dataset}

\item{quantileIntensity}{in percent, how many most intense points are to be kept}

\item{CC_level}{a connected components parameter for fineness in cloudcompare (default: 10)}

\item{CC_numberPoints}{the minimum size in points of a component to not be dropped (default: 1000)}

\item{clipHeight}{in m if you need to set all stem seeds to a uniform height}

\item{bottomCut}{only in combination with clipHeight, in m lower border for cutting of relative heights, default to 0}

\item{bushPreparation}{if TRUE then there is additional density filtering to remove low branches and bushes}

\item{numberIntensity}{if you rather want to specify a number of most intense points to be kept}

\item{silent}{fewer outputs if set to TRUE}

\item{fast}{if TRUE, no detailled per stem information will be calculated}
}
\description{
Takes the vegetation cloud from "extractVegetation" and seperates the stems
according to their intensity values
}
