% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{thMk}
\alias{thMk}
\title{Creates values with thousandmark (are prettier to look at)}
\usage{
thMk(val)
}
\arguments{
\item{val}{Numeric value that is extended with thousand mark 15200000 -> 15.200.000}
}
\description{
Creates values with thousandmark (are prettier to look at)
}
