% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x0_groundRemovalDTM.R
\name{transformVegetation}
\alias{transformVegetation}
\title{Ground classification of raw point cloud with transformation into old set
for more help see function extractVegetation,
which is run twice (local & transformed) within this function}
\usage{
transformVegetation(
  LASfile,
  fileFinder,
  icp.voxel.new = 0,
  icp.voxel.old = 0,
  icp.voxel.both = 0,
  icpMinDist = 1,
  icpIterations = 100,
  icpType = "rigid",
  runInitialExtractVegetation = T,
  groundMergeCut = 0,
  ipad = FALSE,
  groundCutHeight = 1,
  steepSlope = TRUE,
  clothSize = 0.1,
  doFilter = FALSE,
  filterRes = 0.05,
  filterN = 27,
  clip.radius = 0,
  clip.trajectory.distance = 0,
  clip.x = 0,
  clip.y = 0,
  draw.trajectory = TRUE,
  selector = "xyzcit0RGB",
  dtm.path = "",
  preTrafoMatrix.path = "",
  trafo.matchOldSet = "",
  trafo.matchDirPath = "",
  trafo.matchSlice = "_clusterSlice_120to140.laz",
  rainFilter = 0,
  additionalSlices = TRUE,
  exportSlice.upperLimit = 3,
  exportSlice.lowerLimit = 1,
  exportClippedLAS = FALSE,
  dirPath = paste0(getwd(), "/")
)
}
\arguments{
\item{icp.voxel.both}{set the voxel size before slice files are matched via ICP}

\item{icpMinDist}{to Morpho::icpmat() restrict valid points to be within this distance}

\item{icpIterations}{to Morpho::icpmat() number of iterations}

\item{icpType}{to Morpho::icpmat() select the transform to be applied, can be "rigid","similarity" or "affine"}

\item{trafo.matchOldSet}{name of set to try for matching new set into}

\item{trafo.matchDirPath}{directory of the old fileFinder which is used for}

\item{trafo.matchSlice}{can be set to cut a circle from 0 0 with the given radius of the big input file}
}
\description{
Reads in a .las or .laz file from disk and extracts the vegetation cloud
using the cloth simulation filter (scf) via lasground from the lidR package
automatically computing transformation matrix
to merge new set fileFinder into old set trafo.matchOldSet
}
