% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\name{treeCol}
\alias{treeCol}
\title{Convert abbreviation or numeric code to tree species color}
\usage{
treeCol(number, modeNumber = FALSE)
}
\arguments{
\item{number}{abbreviation for tree species with two letters, be for beech, sp for spruce (English to German), or numeric}

\item{modeNumber}{set to TRUE when the input is a number and not an abbreviation}
}
\description{
Convert abbreviation or numeric code to tree species color
}
