% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R,
%   R/treeExtraction_pls_at.R
\name{treeSpecies}
\alias{treeSpecies}
\title{Convert tree species number to abbreviated name}
\usage{
treeSpecies(number, modeNumber = FALSE)

treeSpecies(number, modeNumber = FALSE)
}
\arguments{
\item{number}{abbreviation for tree species with two letters, be for beech, sp for spruce (English to German)}

\item{modeNumber}{set to TRUE when the input is a number and not an abbreviation}
}
\description{
e.g. 1 = "Fi"
}
