% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad1_createAppFiles_tablet.R
\name{treeSpeciesNumber}
\alias{treeSpeciesNumber}
\title{Convert tree species abbreviated name to the inventory number}
\usage{
treeSpeciesNumber(species)
}
\arguments{
\item{species}{German species name that is converted to number for inventory list}
}
\description{
e.g. "Fi" = 1
}
