% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeX-package.R
\docType{package}
\name{treeX-package}
\alias{treeX}
\alias{treeX-package}
\title{treeX: Individual Tree Extraction of Ground-Based Point Clouds}
\description{
Perform tree detection, instance segmentation and automatic measurements of individual trees in LiDAR point clouds.
}
\author{
\strong{Maintainer}: Andreas Tockner \email{andreas.tockner@boku.ac.at} (\href{https://orcid.org/0000-0001-6833-6713}{ORCID})

}
\keyword{internal}
