% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeExtraction_pls_at.R
\docType{data}
\name{v.env}
\alias{v.env}
\title{v.env stores the path _total_ground_veg}
\format{
An object of class \code{environment} of length 2.
}
\usage{
v.env
}
\description{
v.env stores the path _total_ground_veg
}
\keyword{datasets}
